Login

+++ repository/repository.recipes.go
@@ -19,9 +19,8 @@ func NewRecipeRepository(db *recipeDb.RecipeDb) RecipeRepository {
 	}
 }
 
-func (r *RecipeRepository) GetRecipes() (*[]models.Recipe, error) {
-
-	rows, err := r.db.SqlDb.Query("SELECT * FROM recipe")
+func (r *RecipeRepository) GetRecipes(recipeUserId int) (*[]models.Recipe, error) {
+	rows, err := r.db.SqlDb.Query("SELECT * FROM recipe where recipe_user_id=$1", recipeUserId)
 	if err != nil {
 		log.Print(err)
 	}
@@ -33,7 +32,7 @@ func (r *RecipeRepository) GetRecipes() (*[]models.Recipe, error) {
 		var r models.Recipe
 		err := rows.Scan(
 			&r.Id,
-			&r.AccountId,
+			&r.RecipeUserId,
 			&r.RecipeName,
 			&r.RecipeSteps,
 			&r.CreatedOn,
@@ -43,7 +42,6 @@ func (r *RecipeRepository) GetRecipes() (*[]models.Recipe, error) {
 		}
 
 		recipes = append(recipes, r)
-		log.Print(recipes)
 	}
 	err = rows.Err()
 	if err != nil {
@@ -53,14 +51,14 @@ func (r *RecipeRepository) GetRecipes() (*[]models.Recipe, error) {
 	return &recipes, nil
 }
 
-func (r *RecipeRepository) GetRecipe(recipeId int) (*models.Recipe, error) {
+func (r *RecipeRepository) GetRecipe(recipeId int, recipeUserid int) (*models.Recipe, error) {
 
-	row := r.db.SqlDb.QueryRow("SELECT * FROM recipe WHERE id=$1", recipeId)
+	row := r.db.SqlDb.QueryRow("SELECT * FROM recipe WHERE id=$1 AND recipe_user_id=$2", recipeId, recipeUserid)
 	var recipe models.Recipe
 
 	switch err := row.Scan(
 		&recipe.Id,
-		&recipe.AccountId,
+		&recipe.RecipeUserId,
 		&recipe.RecipeName,
 		&recipe.RecipeSteps,
 		&recipe.CreatedOn,
@@ -75,9 +73,9 @@ func (r *RecipeRepository) GetRecipe(recipeId int) (*models.Recipe, error) {
 	}
 }
 
-func (r *RecipeRepository) InsertRecipe(nr *models.Recipe) (b int64, err error) {
+func (r *RecipeRepository) InsertRecipe(recipeUserId int, ir *models.SaveRecipe) (b int64, err error) {
 	var id int64
-	var cols = "(account_id, recipe_name, recipe_steps, created_on, updated_on)"
+	var cols = "(recipe_user_id, recipe_name, recipe_steps, created_on, updated_on)"
 	var values = "($1, $2, $3, now(), now())"
 
 	var query = fmt.Sprintf(
@@ -87,7 +85,7 @@ func (r *RecipeRepository) InsertRecipe(nr *models.Recipe) (b int64, err error) 
 
 	if err := r.db.SqlDb.QueryRow(
 		query,
-		nr.AccountId, nr.RecipeName, nr.RecipeSteps,
+		recipeUserId, ir.RecipeName, ir.RecipeSteps,
 	).Scan(&id); err != nil {
 		panic(err)
 	}
@@ -100,13 +98,13 @@ func (r *RecipeRepository) InsertRecipe(nr *models.Recipe) (b int64, err error) 
 	return id, nil
 }
 
-func (r *RecipeRepository) UpdateRecipe(recipe *models.Recipe, recipeid int) (d bool, err error) {
+func (r *RecipeRepository) UpdateRecipe(recipeid int, recipeUserId int, recipe *models.SaveRecipe) (d bool, err error) {
 	q := `
 		UPDATE recipe
-		SET recipe_name = $2, recipe_steps = $3
-		WHERE id = $1;`
+		SET recipe_name = $3, recipe_steps = $4
+		WHERE id = $1 AND recipe_user_id = $2;`
 
-	_, err = r.db.SqlDb.Exec(q, recipeid, recipe.RecipeName, recipe.RecipeSteps)
+	_, err = r.db.SqlDb.Exec(q, recipeid, recipeUserId, recipe.RecipeName, recipe.RecipeSteps)
 	if err != nil {
 		log.Print(err)
 	}
@@ -114,13 +112,69 @@ func (r *RecipeRepository) UpdateRecipe(recipe *models.Recipe, recipeid int) (d 
 	return true, nil
 }
 
-func (r *RecipeRepository) DeleteRecipe(recipeId int) (d bool, err error) {
-	q := `DELETE FROM recipe WHERE id=$1`
-	_, err = r.db.SqlDb.Exec(q, recipeId)
+func (r *RecipeRepository) DeleteRecipe(recipeId int, recipeUserId int) (d bool, err error) {
+	q := `DELETE FROM recipe WHERE id=$1 AND recipe_user_id=$2;`
+	_, err = r.db.SqlDb.Exec(q, recipeId, recipeUserId)
 
 	if err != nil {
 		log.Print(err)
 	}
 
 	return true, nil
+}
+
+func (r *RecipeRepository) InsertRecipeUser(firstname string, lastname string, email string, hashedPwd string) (b int64, err error) {
+	var id int64
+	var cols = "(first_name, last_name, email, password, created_on, updated_on)"
+	var values = "($1, $2, $3, $4, now(), now())"
+
+	var query = fmt.Sprintf(
+		"INSERT INTO recipe_user %s VALUES %s RETURNING id",
+		cols, values,
+	)
+
+	if err := r.db.SqlDb.QueryRow(
+		query,
+		firstname, lastname, email, hashedPwd,
+	).Scan(&id); err != nil {
+		log.Print(err)
+		return 0, err
+	}
+
+	return id, nil
+}
+
+func (r *RecipeRepository) DeleteRecipeUser(email string) (d bool, err error) {
+	q := "DELETE FROM recipe_user WHERE email=$1;"
+	_, err = r.db.SqlDb.Exec(q, email)
+
+	if err != nil {
+		log.Print(err)
+	}
+
+	return true, nil
+}
+
+func (r *RecipeRepository) GetRecipeUserPwd(email string) (*models.RecipeUser, error) {
+
+	row := r.db.SqlDb.QueryRow("SELECT * FROM recipe_user WHERE email=$1", email)
+
+	var ru models.RecipeUser
+
+	switch err := row.Scan(
+		&ru.Id,
+		&ru.Firstname,
+		&ru.Password,
+		&ru.Email,
+		&ru.Password,
+		&ru.CreatedOn,
+		&ru.UpdatedOn,
+	); err {
+	case sql.ErrNoRows:
+		return nil, nil
+	case nil:
+		return &ru, nil
+	default:
+		panic(err)
+	}
 }
